[Fix] 네트워킹 파일 분리
