+++ HMOA_iOS/HMOA_iOS/Source/Network/ServiceAPI.swift
@@ -19,7 +19,7 @@ enum NetworkError: Error {
 }
 
 
-fileprivate func networking<T: Decodable>(
+public func networking<T: Decodable>(
     urlStr: String,
     method: HTTPMethod,
     data: Data?,
@@ -59,111 +59,4 @@ fileprivate func networking<T: Decodable>(
     }
 }
 
-final class API {
-    
-    /// 향수 한 개 받아오기
-    static func getPerfumeOne(perfumeId: Int) -> Observable<Object> {
-        
-        return networking(
-            urlStr: Address.perfumeOne(perfumeId).url,
-            method: .get,
-            data: nil,
-            model: Object.self)
-    }
-    
-    static func postPerfumeLike(perfumeId: Int) -> Observable<String> {
-        
-        return networking(
-            urlStr: Address.perfumeOne(perfumeId).url,
-            method: .post,
-            data: nil,
-            model: String.self)
-        
-    }
-    
-    //googleToken 보내기
-    static func postAccessToken(params: [String: String]) -> Observable<GoogleToken> {
-        
-        guard let data = try? JSONSerialization.data(
-                    withJSONObject: params,
-                    options: .prettyPrinted)
-                
-        else { return Observable.error(NetworkError.invalidParameters)}
-        
-        return networking(
-            urlStr: Address.postToken.url,
-            method: .post,
-            data: data,
-            model: GoogleToken.self)
-        
-    }
-    
-    //닉네임 중복 검사
-    static func checkDuplicateNickname(params: [String: String]) -> Observable<Bool> {
-        
-        guard let data = try? JSONSerialization.data(withJSONObject: params)
-        else { return Observable.error(NetworkError.invalidParameters) }
 
-        
-        return networking(
-            urlStr: Address.checkNickname.url,
-            method: .post,
-            data: data,
-            model: Bool.self)
-        .map { result -> Bool in
-            return result
-        }
-        .catch { error -> Observable<Bool> in
-            if let statusCode = error.asAFError?.responseCode, statusCode == 409 {
-                return Observable.just(true)
-            }
-            return Observable.error(error)
-        }
-    }
-    
-    //닉네임 업데이트
-    static func updateNickname(params: [String: String]) -> Observable<Response> {
-        
-        guard let data = try? JSONSerialization.data(withJSONObject: params)
-        else { return Observable.error(NetworkError.invalidParameters) }
-        
-        return networking(
-            urlStr: Address.patchNickname.url,
-            method: .patch,
-            data: data,
-            model: Response.self)
-    }
-    
-    static func updateSex(params: [String: String]) -> Observable<Response> {
-        guard let data = try? JSONSerialization.data(withJSONObject: params)
-        else { return Observable.error(NetworkError.invalidParameters) }
-        
-        return networking(
-            urlStr: Address.patchSex.url,
-            method: .patch,
-            data: data,
-            model: Response.self)
-    }
-    
-    static func updateAge(params: [String: Int]) -> Observable<Response> {
-        guard let data = try? JSONSerialization.data(withJSONObject: params)
-        else { return Observable.error(NetworkError.invalidParameters) }
-        
-        return networking(
-            urlStr: Address.patchAge.url,
-            method: .patch,
-            data: data,
-            model: Response.self)
-    }
-    
-    static func join(params: [String: Any]) -> Observable<JoinResponse> {
-        guard let data = try? JSONSerialization.data(withJSONObject: params)
-        else { return Observable.error(NetworkError.invalidParameters) }
-        
-        return networking(
-            urlStr: Address.patchJoin.url,
-            method: .patch,
-            data: data,
-            model: JoinResponse.self)
-    }
-}
